+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 ## unreleased
 
 - Support API 4.9
+- Added poll answer event
 
 ## 0.0.41
 
-- Fix answerPreCheckoutQuery typo
+- Fixed answerPreCheckoutQuery typo
 
 ## 0.0.40